feat: add routing, main, about and 404 pages
