+++ react-app/src/App.tsx
@@ -1,51 +1,26 @@
-import React, { PropsWithChildren } from 'react';
+import { Routes, Route, BrowserRouter } from 'react-router-dom';
+import React from 'react';
+import MainPage from './pages/MainPage/MainPage';
+import AboutPage from './pages/AboutPage/AboutPage';
+import NotFoundPage from './pages/NotFoundPage/NotFoundPage';
 import Header from './components/Header/Header';
-import SearchForm from './components/SearchForm/SearchForm';
-import RepoList from './components/RepoList/RepoList';
 
 
-interface State {
-  currentLanguage: string;
-  currentKeyword: string;
-}
+export default class App extends React.Component<unknown> {
 
-export default class App extends React.Component<PropsWithChildren, State> {
-
-  constructor(props: PropsWithChildren) {
+  constructor(props: unknown) {
     super(props);
-    this.state = {
-      currentLanguage: localStorage.getItem('currentLanguage') || 'javascript',
-      currentKeyword: localStorage.getItem('currentKeyword') || '',
-    };
-    this.setCurrentQueryParams = this.setCurrentQueryParams.bind(this);
-  }
-
-
-  setCurrentQueryParams(params: { language: string, keyword: string }) {
-    const { language, keyword } = params;
-    this.setState(() => ({
-      currentLanguage: language,
-      currentKeyword: keyword,
-    }));
-    localStorage.setItem('currentLanguage', language);
-    localStorage.setItem('currentKeyword', keyword);
   }
 
   render() {
-    const { currentLanguage, currentKeyword } = this.state;
     return (
-      <div className="lg:container px-5 py-24 mx-auto text-gray-900">
-        <Header>
-          <SearchForm
-            setCurrentLanguage={this.setCurrentQueryParams}
-            currentLanguage={currentLanguage}
-            currentKeyword={currentKeyword}
-          />
-        </Header>
-        <main>
-          <RepoList currentLanguage={currentLanguage} currentKeyword={currentKeyword}/>
-        </main>
-      </div>
-    );
+      <BrowserRouter basename={import.meta.env.BASE_URL}>
+        <Header/>
+        <Routes>
+          <Route path="/" element={<MainPage/>}/>
+          <Route path="/about" element={<AboutPage/>}/>
+          <Route path="/*" element={<NotFoundPage/>}/>
+        </Routes>
+      </BrowserRouter>);
   }
 }
