+++ react-app/src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default class NotFoundPage extends React.Component {
+  render() {
+    return (<p className="text-center">There is nothing here...</p>);
+  }
+}
